BREAKING CHANGE Change biquad coefficient from float[5] to double[5] while stored.
This makes no adverse speed change due to saved conversions.
Sonically THIS MAKES A SMALL AUDIBLE DIFFERENCE of up tp -50 dB RMS relatively to the previous float[5]
But there is little reason to worry, as it all means the filters will be more in line with their specifications now.

+++ dsp/dplug/dsp/iir.d
@@ -31,7 +31,7 @@ public
     ///       coeff[2] is b2,
     ///       coeff[3] is a1,
     ///       coeff[4] is a2 in the litterature.
-    alias BiquadCoeff = float[5];
+    alias BiquadCoeff = double[5];
 
     /// Maintain state for a biquad state.
     /// A biquad is a realization that can model two poles and two zeros.
